Fixed several warnings appeared with recent GCC versions
* improved printf format checks.
* fixed formatting weaknesses
* fixed "unmanaged results" in some functions.
* fixed unused paramerers
* tools: adds missing 'import syslog' in monitoring scripts.

+++ rawx-client/lib/stats.c
@@ -83,7 +83,7 @@ rawx_client_get_statistics(rawx_session_t * session, GError ** err)
 	case NE_AUTH:
 		str_addr_size = addr_info_to_string(&(session->addr), str_addr, sizeof(str_addr));
 		GSETERROR(err, "cannot download the stats from [%.*s]' (%s)",
-		    str_addr_size, str_addr, ne_get_error(session->neon_session));
+		    (int)str_addr_size, str_addr, ne_get_error(session->neon_session));
 		break;
 	default:
 		GSETERROR(err, "Unexpected return code from the neon library : %d", rc);
